--- a/internal/runner/context/exec_info.go
+++ b/internal/runner/context/exec_info.go
@@ -2,9 +2,9 @@ package runner
 
 import "context"
 
-type runnerContextKey struct{}
+type contextKey struct{ string }
 
-var ExecutionInfoKey = &runnerContextKey{}
+var executionInfoKey = &contextKey{"ExecutionInfo"}
 
 type ExecutionInfo struct {
 	ExecContext string
@@ -13,6 +13,11 @@ type ExecutionInfo struct {
 	RunID       string
 }
 
-func ContextWithExecutionInfo(ctx context.Context, execInfo *ExecutionInfo) context.Context {
-	return context.WithValue(ctx, ExecutionInfoKey, execInfo)
+func WithExecutionInfo(ctx context.Context, execInfo *ExecutionInfo) context.Context {
+	return context.WithValue(ctx, executionInfoKey, execInfo)
+}
+
+func ExecutionInfoFromContext(ctx context.Context) (*ExecutionInfo, bool) {
+	execInfo, ok := ctx.Value(executionInfoKey).(*ExecutionInfo)
+	return execInfo, ok
 }