--- a/internal/runnerv2service/service_sessions.go
+++ b/internal/runnerv2service/service_sessions.go
@@ -11,31 +11,7 @@ import (
 	rcontext "github.com/stateful/runme/v3/internal/runner/context"
 	"github.com/stateful/runme/v3/internal/session"
 	runnerv2 "github.com/stateful/runme/v3/pkg/api/gen/proto/go/runme/runner/v2"
-	"github.com/stateful/runme/v3/pkg/project"
 )
-
-func convertSessionToRunnerv2alpha1Session(sess *session.Session) *runnerv2.Session {
-	return &runnerv2.Session{
-		Id:  sess.ID,
-		Env: sess.GetAllEnv(),
-		// Metadata: sess.Metadata,
-	}
-}
-
-// TODO(adamb): this function should not return nil project and nil error at the same time.
-func convertProtoProjectToProject(runnerProj *runnerv2.Project) (*project.Project, error) {
-	if runnerProj == nil {
-		return nil, nil
-	}
-
-	opts := project.DefaultProjectOptions[:]
-
-	if runnerProj.EnvLoadOrder != nil {
-		opts = append(opts, project.WithEnvFilesReadOrder(runnerProj.EnvLoadOrder))
-	}
-
-	return project.NewDirProject(runnerProj.Root, opts...)
-}
 
 func (r *runnerService) CreateSession(ctx context.Context, req *runnerv2.CreateSessionRequest) (*runnerv2.CreateSessionResponse, error) {
 	r.logger.Info("running CreateSession in runnerService")
@@ -75,7 +51,7 @@ func (r *runnerService) CreateSession(ctx context.Context, req *runnerv2.CreateS
 	r.logger.Debug("created session", zap.String("id", sess.ID), zap.Bool("owl", owl), zap.Int("seed_env_len", len(seedEnv)))
 
 	return &runnerv2.CreateSessionResponse{
-		Session: convertSessionToRunnerv2alpha1Session(sess),
+		Session: convertSessionToProtoSession(sess),
 	}, nil
 }
 
@@ -88,7 +64,7 @@ func (r *runnerService) GetSession(_ context.Context, req *runnerv2.GetSessionRe
 	}
 
 	return &runnerv2.GetSessionResponse{
-		Session: convertSessionToRunnerv2alpha1Session(sess),
+		Session: convertSessionToProtoSession(sess),
 	}, nil
 }
 
@@ -99,7 +75,7 @@ func (r *runnerService) ListSessions(_ context.Context, req *runnerv2.ListSessio
 
 	runnerSessions := make([]*runnerv2.Session, 0, len(sessions))
 	for _, s := range sessions {
-		runnerSessions = append(runnerSessions, convertSessionToRunnerv2alpha1Session(s))
+		runnerSessions = append(runnerSessions, convertSessionToProtoSession(s))
 	}
 
 	return &runnerv2.ListSessionsResponse{Sessions: runnerSessions}, nil
@@ -117,7 +93,7 @@ func (r *runnerService) UpdateSession(ctx context.Context, req *runnerv2.UpdateS
 		return nil, err
 	}
 
-	return &runnerv2.UpdateSessionResponse{Session: convertSessionToRunnerv2alpha1Session(sess)}, nil
+	return &runnerv2.UpdateSessionResponse{Session: convertSessionToProtoSession(sess)}, nil
 }
 
 func (r *runnerService) DeleteSession(_ context.Context, req *runnerv2.DeleteSessionRequest) (*runnerv2.DeleteSessionResponse, error) {
@@ -138,7 +114,7 @@ type updateRequest interface {
 }
 
 func (r *runnerService) updateSession(ctx context.Context, sess *session.Session, req updateRequest) error {
-	ctx = rcontext.ContextWithExecutionInfo(ctx, &rcontext.ExecutionInfo{
+	ctx = rcontext.WithExecutionInfo(ctx, &rcontext.ExecutionInfo{
 		ExecContext: "request",
 	})