This PR refactors the `runnerv2service` package to speed up the performance of executing programs producing large output.

### Test

`cat` a large file and pipe the output to `/dev/null`. As the execution goes through the server, the execution involves marshaling and unmarshaling to Protobuf.

#### This PR

```
time ./runme beta run --remote generate | cat - >/dev/null
./runme beta run --remote generate  0.70s user 0.64s system 33% cpu 3.973 total
cat - > /dev/null  0.01s user 0.06s system 1% cpu 3.972 total
```

#### Main

```
time ./runme beta run --remote generate | cat - >/dev/null
./runme beta run --remote generate  5.58s user 6.89s system 146% cpu 8.537 total
cat - > /dev/null  0.21s user 1.04s system 14% cpu 8.536 total
```

We have a speed increase of over 2x.